refactor: extract ReviewWorkspaceTypes.swift from ReviewWorkspace.swift
Extracts 4 support types (ReviewWorkspaceTab, ReviewWorkspaceTimelineEntry,
ReviewWorkspaceTimelineEntryKind, ReviewSummaryIssueGroup) from
ReviewWorkspace.swift into a new ReviewWorkspaceTypes.swift.

Byte-identical move; SHA-256 c1f214d8ae2a97b63b8dcb693f03f28bfd978170ba77841ebfc7b618b98ab10b
verified against origin/main:ReviewWorkspace.swift lines 143-213.

Reduces ReviewWorkspace.swift from 213 to 141 lines. Only the primary
ReviewWorkspace enum remains.

Tests: 667 (unchanged).
